[RENAME] 상위 폴더로 이동 및 불필요한 파일 삭제
- FooterNav.tsx 파일 Footer 폴더로 이동
- FooterNav 폴더 및 FooterNav.module.css 파일 삭제
