--- src/components/Footer/FooterNav/FooterNav.module.css
@@ -1,14 +1,0 @@
-.footerNav {
-
-  p {
-    font-weight: 600;
-    padding-bottom: 1.5rem;
-    font-size: var(--large 1.25rem);
-  }
-
-  a {
-    font-size: var(--medium 1rem);
-    line-height: 200%;
-    text-decoration: none;
-  }
-}
